Initialization and UI adjustments:
- fine tuning for start/stop database connections and derby server
- adjusted table cell colors for different themes for web settings

+++ eclipseProjects/de.enflexit.awb/bundles/database/de.enflexit.db.derby/src/de/enflexit/db/derby/server/DerbyBundleActivator.java
@@ -2,6 +2,7 @@ package de.enflexit.db.derby.server;
 
 import java.nio.file.Path;
 
+import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleActivator;
 import org.osgi.framework.BundleContext;
 
@@ -32,12 +33,47 @@ public class DerbyBundleActivator implements BundleActivator {
 		DerbyNetworkServer.execute();
 	}
 	
+	
 	/* (non-Javadoc)
 	 * @see org.osgi.framework.BundleActivator#stop(org.osgi.framework.BundleContext)
 	 */
 	@Override
 	public void stop(BundleContext context) throws Exception {
-		DerbyNetworkServer.terminate();
+		
+		// --- Shutdown the database server? ----------------------------------
+		if (DerbyNetworkServer.isExecuted()==true) {
+			// --- Ensure that the Hibernate bundle is stopped before --------- 
+			try {
+				Bundle hibernateBundle = this.getBundleByName(context.getBundles(), "de.enflexit.db.hibernate");
+				if (hibernateBundle!=null) {
+					int dState = hibernateBundle.getState();
+					if (dState==Bundle.ACTIVE || dState==Bundle.STOPPING || dState==Bundle.STOP_TRANSIENT) {
+						hibernateBundle.stop();
+					}
+				}
+				
+			} catch (Exception ex) {
+				ex.printStackTrace();
+			}
+			
+			// --- Terminate Derby database server ----------------------------
+			DerbyNetworkServer.terminate();
+		}
 	}
-
+	/**
+	 * If to be found, returns the bundle instance for the specified bundle name.
+	 *
+	 * @param bundles the bundles
+	 * @param symBundleName the sym bundle name
+	 * @return the bundle by name
+	 */
+	private Bundle getBundleByName(Bundle[] bundles, String symBundleName) {
+		for (Bundle bundle : bundles) {
+			if (bundle.getSymbolicName().equals(symBundleName)==true) {
+				return bundle;
+			}
+		}
+		return null;
+	}
+	
 }
